+++ Dmtd.Module/ViewModels/PhaseLogExportScope.cs
@@ -1,0 +1,8 @@
+namespace Dmtd.Module.ViewModels;
+
+public enum PhaseLogExportScope
+{
+    Session,
+    RetentionPeriod,
+    CustomRange
+}